+++ node/test/inputtests.ts
@@ -30,7 +30,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['INPUT_UNITTESTINPUT'] = 'test value';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var inval = tl.getInput('UnitTestInput', true);
         assert.equal(inval, 'test value');
@@ -41,7 +41,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['INPUT_UNITTESTINPUT'] = 'test value';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         var inval = tl.getInput('UnitTestInput', true);
         assert.equal(inval, 'test value');
         assert(!process.env['INPUT_UNITTESTINPUT'], 'input envvar should be cleared');
@@ -66,7 +66,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['INPUT_UNITTESTINPUT'] = '   test value   ';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var inval = tl.getInput('UnitTestInput', true);
         assert.equal(inval, 'test value');
@@ -79,7 +79,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['BUILD_REPOSITORY_NAME'] = 'Test Repository';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var varVal = tl.getVariable('Build.Repository.Name');
         assert.equal(varVal, 'Test Repository');
@@ -90,7 +90,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['SECRET_BUILD_REPOSITORY_NAME'] = 'Test Repository';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var varVal = tl.getVariable('Build.Repository.Name');
         assert.equal(varVal, 'Test Repository');
@@ -102,7 +102,7 @@ describe('Input Tests', function () {
 
         process.env['BUILD_REPOSITORY_NAME'] = 'Test Repository';
         process.env['SECRET_BUILD_REPOSITORY_NAME'] = 'Secret Test Repository';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var varVal = tl.getVariable('Build.Repository.Name');
         assert.equal(varVal, 'Secret Test Repository');
@@ -161,7 +161,7 @@ describe('Input Tests', function () {
     it('does not allow a secret variable to become a public variable', function (done) {
         this.timeout(1000);
 
-        tl._loadData();
+        (tl as any)._internal._loadData();
         tl.setVariable('My.Secret.Var', 'test secret value', true);
         tl.setVariable('My.Secret.Var', 'test modified value', false);
         let vars: tl.VariableInfo[] = tl.getVariables();
@@ -175,7 +175,7 @@ describe('Input Tests', function () {
     it('allows a public variable to become a secret variable', function (done) {
         this.timeout(1000);
 
-        tl._loadData();
+        (tl as any)._internal._loadData();
         tl.setVariable('My.Var', 'test value', false);
         tl.setVariable('My.Var', 'test modified value', true);
         let vars: tl.VariableInfo[] = tl.getVariables();
@@ -189,7 +189,7 @@ describe('Input Tests', function () {
     it('tracks known variables using env formatted name', function (done) {
         this.timeout(1000);
 
-        tl._loadData();
+        (tl as any)._internal._loadData();
         tl.setVariable('My.Public.Var', 'test value');
         tl.setVariable('my_public.VAR', 'test modified value');
         let vars: tl.VariableInfo[] = tl.getVariables();
@@ -207,7 +207,7 @@ describe('Input Tests', function () {
         process.env['PUBLIC_VAR_ONE'] = 'public value 1';
         process.env['PUBLIC_VAR_TWO'] = 'public value 2';
         process.env['VSTS_PUBLIC_VARIABLES'] = '[ "Public.Var.One", "Public.Var.Two" ]';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         let vars: tl.VariableInfo[] = tl.getVariables();
         assert(vars, 'variables should not be undefined or null');
         assert.equal(vars.length, 2, 'exactly 2 variables should be returned');
@@ -227,7 +227,7 @@ describe('Input Tests', function () {
         process.env['SECRET_SECRET_VAR_ONE'] = 'secret value 1';
         process.env['SECRET_SECRET_VAR_TWO'] = 'secret value 2';
         process.env['VSTS_SECRET_VARIABLES'] = '[ "Secret.Var.One", "Secret.Var.Two" ]';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         let vars: tl.VariableInfo[] = tl.getVariables();
         assert(vars, 'variables should not be undefined or null');
         assert.equal(vars.length, 2, 'exactly 2 variables should be returned');
@@ -246,7 +246,7 @@ describe('Input Tests', function () {
 
         process.env['SECRET_SECRET_VAR_ONE'] = 'secret value 1';
         process.env['SECRET_SECRET_VAR_TWO'] = 'secret value 2';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         let vars: tl.VariableInfo[] = tl.getVariables();
         assert(vars, 'variables should not be undefined or null');
         assert.equal(vars.length, 2, 'exactly 2 variables should be returned');
@@ -265,7 +265,7 @@ describe('Input Tests', function () {
 
         process.env['INITIAL_PUBLIC_VAR'] = 'initial public value';
         process.env['VSTS_PUBLIC_VARIABLES'] = '[ "Initial.Public.Var" ]';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         tl.setVariable('Set.Public.Var', 'set public value');
         let vars: tl.VariableInfo[] = tl.getVariables();
         assert(vars, 'variables should not be undefined or null');
@@ -285,7 +285,7 @@ describe('Input Tests', function () {
 
         process.env['SECRET_INITIAL_SECRET_VAR'] = 'initial secret value';
         process.env['VSTS_SECRET_VARIABLES'] = '[ "Initial.Secret.Var" ]';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         tl.setVariable('Set.Secret.Var', 'set secret value', true);
         let vars: tl.VariableInfo[] = tl.getVariables();
         assert(vars, 'variables should not be undefined or null');
@@ -306,7 +306,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['ENDPOINT_URL_id1'] = 'http://url';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var url = tl.getEndpointUrl('id1', true);
         assert.equal(url, 'http://url', 'url should match');
@@ -317,7 +317,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['ENDPOINT_AUTH_id1'] = '{ "parameters": {"param1": "val1", "param2": "val2"}, "scheme": "UsernamePassword"}';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var auth = tl.getEndpointAuthorization('id1', true);
         assert(auth, 'should return an auth obj');
@@ -329,7 +329,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         // don't set
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var auth = tl.getEndpointAuthorization('id1', true);
         assert.equal(auth, null, 'should not return an auth obj');
@@ -340,7 +340,7 @@ describe('Input Tests', function () {
         this.timeout(1000);
 
         process.env['ENDPOINT_AUTH_id1'] = '{ "parameters": {"param1": "val1", "param2": "val2"}, "scheme": "UsernamePassword"}';
-        tl._loadData();
+        (tl as any)._internal._loadData();
         var auth = tl.getEndpointAuthorization('id1', true);
         assert(auth, 'should return an auth obj');
         assert(auth['parameters']['param1'] === 'val1', 'should be correct object');
@@ -351,7 +351,7 @@ describe('Input Tests', function () {
     it('gets endpoint auth scheme', function (done) {
         this.timeout(1000);
         process.env['ENDPOINT_AUTH_SCHEME_id1'] = 'scheme1';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var data = tl.getEndpointAuthorizationScheme('id1', true);
         assert(data, 'should return a string value');
@@ -362,7 +362,7 @@ describe('Input Tests', function () {
     })
     it('gets undefined if endpoint auth scheme is not set', function (done) {
         this.timeout(1000);
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var data = tl.getEndpointAuthorizationScheme('id1', true);
         assert(!data, 'should be undefined when auth scheme is not set');
@@ -372,7 +372,7 @@ describe('Input Tests', function () {
     it('gets endpoint auth parameters', function (done) {
         this.timeout(1000);
         process.env['ENDPOINT_AUTH_PARAMETER_id1_PARAM1'] = 'value1';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var data = tl.getEndpointAuthorizationParameter('id1', 'param1', true);
         assert(data, 'should return a string value');
@@ -383,7 +383,7 @@ describe('Input Tests', function () {
     })
     it('gets undefined if endpoint auth parameter is not set', function (done) {
         this.timeout(1000);
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var data = tl.getEndpointAuthorizationParameter('id1', 'noparam', true);
         assert(!data, 'should be undefined when auth param is not set');
@@ -393,7 +393,7 @@ describe('Input Tests', function () {
     it('gets an endpoint data', function (done) {
         this.timeout(1000);
         process.env['ENDPOINT_DATA_id1_PARAM1'] = 'val1';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var data = tl.getEndpointDataParameter('id1', 'param1', true);
         assert(data, 'should return a string value');
@@ -403,7 +403,7 @@ describe('Input Tests', function () {
     })
     it('gets undefined if endpoint data is not set', function (done) {
         this.timeout(1000);
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var data = tl.getEndpointDataParameter('id1', 'noparam', true);
         assert.equal(data, undefined, 'Error should occur if endpoint data is not set');
@@ -417,7 +417,7 @@ describe('Input Tests', function () {
 
         var inputValue = 'tRuE';
         process.env['INPUT_ABOOL'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var outVal = tl.getBoolInput('abool', /*required=*/true);
         assert(outVal, 'should return true');
@@ -429,7 +429,7 @@ describe('Input Tests', function () {
 
         var inputValue = 'false';
         process.env['INPUT_ABOOL'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var outVal = tl.getBoolInput('abool', /*required=*/true);
         assert(!outVal, 'should return false');
@@ -439,7 +439,7 @@ describe('Input Tests', function () {
     it('defaults to false bool input value', function(done) {
         this.timeout(1000);
 
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var outVal = tl.getBoolInput('no_such_env_var', /*required=*/ false);
         assert(!outVal, 'should default to false');
@@ -453,7 +453,7 @@ describe('Input Tests', function () {
 
         var inputValue = 'test  value'; // contains two spaces
         process.env['INPUT_DELIM'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var outVal = tl.getDelimitedInput('delim', ' ', /*required*/true);
         assert.equal(outVal.length, 2, 'should return array with two elements');
@@ -467,7 +467,7 @@ describe('Input Tests', function () {
 
         var inputValue = 'testvalue';
         process.env['INPUT_DELIM'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var outVal = tl.getDelimitedInput('delim', ' ', /*required*/true);
         assert.equal(outVal.length, 1, 'should return array with one element');
@@ -480,7 +480,7 @@ describe('Input Tests', function () {
 
         var inputValue = '';
         process.env['INPUT_DELIM'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var outVal = tl.getDelimitedInput('delim', ' ', /*required*/false);
         assert.equal(outVal.length, 0, 'should return array with zero elements');
@@ -494,7 +494,7 @@ describe('Input Tests', function () {
 
         var inputValue = 'test.txt'
         process.env['INPUT_PATH1'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var path = tl.getPathInput('path1', /*required=*/true, /*check=*/false);
         assert(path, 'should return a path');
@@ -556,7 +556,7 @@ describe('Input Tests', function () {
         var inputValue = 'file name.txt';
         var expectedValue = 'file name.txt';
         process.env['INPUT_PATH1'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var path = tl.getPathInput('path1', /*required=*/true, /*check=*/false);
         assert(path, 'should return a path');
@@ -572,7 +572,7 @@ describe('Input Tests', function () {
 
         var inputValue = __filename;
         process.env['INPUT_PATH1'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var path = tl.getPathInput('path1', /*required=*/true, /*check=*/true);
         assert(path, 'should return a path');
@@ -591,7 +591,7 @@ describe('Input Tests', function () {
 
         var inputValue = "someRandomFile.txt";
         process.env['INPUT_PATH1'] = inputValue;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         var worked: boolean = false;
         try {
@@ -612,7 +612,7 @@ describe('Input Tests', function () {
 
         var repoRoot = '/repo/root/dir';
         process.env['INPUT_PATH1'] = repoRoot;
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         process.env['BUILD_SOURCESDIRECTORY'] = repoRoot;
         var supplied = tl.filePathSupplied('path1');
@@ -624,7 +624,7 @@ describe('Input Tests', function () {
 
         var repoRoot = '/repo/root/dir';
         process.env['INPUT_PATH1'] = repoRoot + '/some/path';
-        tl._loadData();
+        (tl as any)._internal._loadData();
 
         process.env['BUILD_SOURCESDIRECTORY'] = repoRoot;
         var supplied = tl.filePathSupplied('path1');